Rearrange src dir to make more sense
