+++ acarshub-typescript/src/helpers/tooltips.ts
